Refactor : memberCommandContoller, QueryController 리팩토링 - #71
