+++ src/main/java/talkPick/domain/member/adapter/in/MemberQueryController.java
@@ -19,6 +19,7 @@ import talkPick.global.annotation.UserId;
 import java.io.IOException;
 import java.util.List;
 
+@RestController
 @RequiredArgsConstructor
 @Slf4j
 @RequestMapping("/api/v1")
