+++ src/main/java/talkPick/domain/member/application/MemberCommandService.java
@@ -1,41 +1,51 @@
 package talkPick.domain.member.application;
 
+import jakarta.persistence.PersistenceException;
 import lombok.RequiredArgsConstructor;
+import org.springframework.dao.DataIntegrityViolationException;
+import org.springframework.orm.jpa.JpaSystemException;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+import talkPick.domain.member.adapter.in.dto.MemberEmailReqDTO;
 import talkPick.domain.member.adapter.out.repository.MemberJpaRepository;
 import talkPick.domain.member.domain.Member;
+import talkPick.domain.member.port.in.MemberCommandUseCase;
+import talkPick.global.error.exception.member.MemberServiceException;
 
 import java.util.Optional;
+
+import static talkPick.domain.member.application.mapper.MemberMapper.fromDtoToMember;
 
 @Service
 @Transactional(readOnly = true)
 @RequiredArgsConstructor
-public class MemberCommandService {
+public class MemberCommandService implements MemberCommandUseCase {
     private final MemberJpaRepository memberJpaRepository;
 
-    public Optional<Member> findByKakaoId(String kakaoId) {
-        if (kakaoId == null || kakaoId.trim().isEmpty()) {
-            throw new IllegalArgumentException("카카오 ID는 비어있을 수 없습니다.");
-        }
+    @Transactional
+    @Override
+    public void setEmailMember(MemberEmailReqDTO memberReqDto) {
+        Member saveMember = fromDtoToMember(memberReqDto);
 
-        return memberJpaRepository.findByKakaoId((kakaoId));
+        memberJpaRepository.save(saveMember);
     }
 
-    public Optional<Member> findById(Long id) {
-        if (id == null) {
-            throw new IllegalArgumentException("회원 ID는 null일 수 없습니다.");
+    @Transactional
+    @Override
+    public Member setKakaoMember(Member member) {
+        try {
+            return memberJpaRepository.save(member);
+        } catch (DataIntegrityViolationException e) {
+            // 데이터 무결성 위반 예외 (예: 중복 키, 제약 조건 위반 등)
+            throw new MemberServiceException("회원 또는 프로필 저장 중 데이터 무결성 오류가 발생했습니다.", e);
+        } catch (JpaSystemException | PersistenceException e) {
+            // JPA 관련 시스템 예외
+            throw new MemberServiceException("JPA 처리 중 오류가 발생했습니다.", e);
+        } catch (Exception e) {
+            // 기타 예외
+            throw new MemberServiceException("회원 및 프로필 저장 중 오류가 발생했습니다.", e);
         }
 
-        return memberJpaRepository.findById(id);
-    }
-
-    public Optional<Member> findByEmail(String email) {
-        if (email == null || email.trim().isEmpty()) {
-            throw new IllegalArgumentException("이메일은 비어있을 수 없습니다.");
-        }
-
-        return memberJpaRepository.findByEmail(email);
     }
